+++ drivers/include/mtd.h
@@ -59,7 +59,7 @@ typedef struct {
     const mtd_desc_t *driver;  /**< MTD driver */
     uint32_t sector_count;     /**< Number of sector in the MTD */
     uint32_t pages_per_sector; /**< Number of pages by sector in the MTD */
-    uint32_t page_size;        /**< Size of the pages in the MTD */
+    size_t page_size;          /**< Size of the pages in the MTD */
 } mtd_dev_t;
 
 /**
@@ -99,7 +99,7 @@ struct mtd_desc {
     int (*read)(mtd_dev_t *dev,
                 void *buff,
                 uint32_t addr,
-                uint32_t size);
+                size_t size);
 
     /**
      * @brief   Write to the Memory Technology Device (MTD)
@@ -118,7 +118,7 @@ struct mtd_desc {
     int (*write)(mtd_dev_t *dev,
                  const void *buff,
                  uint32_t addr,
-                 uint32_t size);
+                 size_t size);
 
     /**
      * @brief   Erase sector(s) over the Memory Technology Device (MTD)
@@ -134,7 +134,7 @@ struct mtd_desc {
      */
     int (*erase)(mtd_dev_t *dev,
                  uint32_t addr,
-                 uint32_t size);
+                 size_t size);
 
     /**
      * @brief   Control power of Memory Technology Device (MTD)
@@ -174,7 +174,7 @@ int mtd_init(mtd_dev_t *mtd);
  * @return -EOVERFLOW if @p addr or @p count are not valid, i.e. outside memory
  * @return -EIO if I/O error occurred
  */
-int mtd_read(mtd_dev_t *mtd, void *dest, uint32_t addr, uint32_t count);
+int mtd_read(mtd_dev_t *mtd, void *dest, uint32_t addr, size_t count);
 
 /**
  * @brief   Write data to a MTD device
@@ -197,7 +197,7 @@ int mtd_read(mtd_dev_t *mtd, void *dest, uint32_t addr, uint32_t count);
  * @return -EIO if I/O error occurred
  * @return -EINVAL if parameters are invalid (invalid alignment for instance)
  */
-int mtd_write(mtd_dev_t *mtd, const void *src, uint32_t addr, uint32_t count);
+int mtd_write(mtd_dev_t *mtd, const void *src, uint32_t addr, size_t count);
 
 /**
  * @brief   Erase sectors of a MTD device
@@ -215,7 +215,7 @@ int mtd_write(mtd_dev_t *mtd, const void *src, uint32_t addr, uint32_t count);
  * @return -EOVERFLOW if @p addr or @p count are not valid, i.e. outside memory
  * @return -EIO if I/O error occurred
  */
-int mtd_erase(mtd_dev_t *mtd, uint32_t addr, uint32_t count);
+int mtd_erase(mtd_dev_t *mtd, uint32_t addr, size_t count);
 
 /**
  * @brief   Set power mode on a MTD device
